requirements.txt
tensorflow>=2.18.0
numpy>=1.26.4
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
pandas>=2.2.0
Pillow>=10.0.0
tqdm>=4.66.0
streamlit>=1.35.0
jupyter>=1.0.0
ipykernel>=6.29.0
